+++ parley-network/src/main/java/nu/parleynetwork/android/data/repository/ParleyNetworkRepositories.java
@@ -3,10 +3,12 @@ package nu.parleynetwork.android.data.repository;
 import nu.parley.android.data.net.ParleyRepositories;
 import nu.parley.android.data.repository.DeviceRepository;
 import nu.parley.android.data.repository.EventRepository;
+import nu.parley.android.data.repository.MessageRepository;
 
 public class ParleyNetworkRepositories implements ParleyRepositories {
     private DeviceRepository deviceRepository = new DeviceRepositoryImpl();
     private EventRepository eventRepository = new EventRepositoryImpl();
+    private MessageRepository messageRepository = new MessageRepositoryImpl();
 
     @Override
     public DeviceRepository getDeviceRepository() {
@@ -16,5 +18,10 @@ public class ParleyNetworkRepositories implements ParleyRepositories {
     @Override
     public EventRepository getEventRepository() {
         return eventRepository;
+    }
+
+    @Override
+    public MessageRepository getMessageRepository() {
+        return messageRepository;
     }
 }
